Created JsonService and transferred JSON related functions there
