+++ BackendFilmes.Service/Interfaces/IMovieService.cs
@@ -9,7 +9,6 @@ namespace BackendFilmes.Service.Interfaces
     {
         public Task<List<Movie>> GetMoviePage(int? page, int? pageSize);
         public List<string> ParseAdditionalParams(string additionalParams);
-        public string MakeJsonWithAdditionalFields(List<Movie> moviesList, List<string> additionalParametersList, int page, int totalPages);
         public int GetTotalPages(int pageSize);
     }
 }
